refactor(generate_types): modularize code by separating into dedicated files
