+++ bin/src/generate_types/default_values.dart
@@ -1,0 +1,8 @@
+/// Default option values
+const defaultValues = {
+  'env': '.env',
+  'output': 'supabase/types',
+  'tag': '',
+  'debug': false,
+  'skipFooter': false,
+};
