+++ bin/src/src.dart
@@ -1,4 +1,5 @@
 export 'generate_supabase_types.dart';
+export 'generate_types/generate_types.dart';
 export 'generators/generators.dart';
 export 'get_types.dart';
 export 'logger.dart';
